Refactor: Move SBType to ClientAPI
