+++ addons/docs/src/frameworks/react/react-argtypes.stories.tsx
@@ -60,7 +60,7 @@ const typescriptFixtures = [
 const typescriptStories = storiesOf('ArgTypes/TypeScript', module);
 typescriptFixtures.forEach((fixture) => {
   // eslint-disable-next-line import/no-dynamic-require, global-require, no-shadow
-  const { Component } = require(`../../lib/sbtypes/__testfixtures__/typescript/${fixture}`);
+  const { Component } = require(`../../lib/convert/__testfixtures__/typescript/${fixture}`);
   typescriptStories.add(fixture, () => <ArgsStory component={Component} />);
 });
 
@@ -69,7 +69,7 @@ const proptypesFixtures = ['arrays', 'enums', 'misc', 'objects', 'react', 'scala
 const proptypesStories = storiesOf('ArgTypes/PropTypes', module);
 proptypesFixtures.forEach((fixture) => {
   // eslint-disable-next-line import/no-dynamic-require, global-require, no-shadow
-  const { Component } = require(`../../lib/sbtypes/__testfixtures__/proptypes/${fixture}`);
+  const { Component } = require(`../../lib/convert/__testfixtures__/proptypes/${fixture}`);
   proptypesStories.add(fixture, () => <ArgsStory component={Component} />);
 });
