+++ addons/docs/src/lib/convert/__testfixtures__/typescript/interfaces.tsx
