+++ addons/docs/src/lib/convert/proptypes/convert.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-case-declarations */
 import mapValues from 'lodash/mapValues';
+import { SBType } from '@storybook/client-api';
 import { PTType } from './types';
-import { SBType } from '../types';
 import { trimQuotes } from '../utils';
 
 const SIGNATURE_REGEXP = /^\(.*\) => /;
