--- addons/docs/src/lib/sbtypes/typescript/index.ts
@@ -1,2 +1,0 @@
-export * from './convert';
-export * from './types';
