+++ lib/client-api/src/types.ts
@@ -104,3 +104,46 @@ export type RenderContext = StoreItem & {
   showError: (error: { title: string; description: string }) => void;
   showException: (err: Error) => void;
 };
+
+interface SBBaseType {
+  required?: boolean;
+  raw?: string;
+}
+
+export type SBScalarType = SBBaseType & {
+  name: 'boolean' | 'string' | 'number' | 'function';
+};
+
+export type SBArrayType = SBBaseType & {
+  name: 'array';
+  value: SBType;
+};
+export type SBObjectType = SBBaseType & {
+  name: 'object';
+  value: Record<string, SBType>;
+};
+export type SBEnumType = SBBaseType & {
+  name: 'enum';
+  value: (string | number)[];
+};
+export type SBIntersectionType = SBBaseType & {
+  name: 'intersection';
+  value: SBType[];
+};
+export type SBUnionType = SBBaseType & {
+  name: 'union';
+  value: SBType[];
+};
+export type SBOtherType = SBBaseType & {
+  name: 'other';
+  value: string;
+};
+
+export type SBType =
+  | SBScalarType
+  | SBEnumType
+  | SBArrayType
+  | SBObjectType
+  | SBIntersectionType
+  | SBUnionType
+  | SBOtherType;
